v3_gemini/requirements.txt
streamlit
langchain
langchain-google-genai
langchain-core
requests
pytz
python-dotenv
